change isbn datatype to BIGINT

+++ server/database/models/book.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, DataTypes) => {
   const Book = sequelize.define('Book', {
     title: DataTypes.STRING,
     author: DataTypes.STRING,
-    isbn: DataTypes.INTEGER,
+    isbn: DataTypes.BIGINT,
   }, {});
   Book.associate = function(models) {
     // associations can be defined here
